requirements.txt
torch==2.7.0
ctranslate2>=4.6.3
faster-whisper
pyaudiowpatch
numpy
pyperclip
requests
dspy-ai
scipy
pydantic>=2.0.0
pydantic-settings>=2.0.0
